+++ R/data.r
@@ -22,6 +22,17 @@
 #'   - **`multipart`**: logical; whether the feature is single or multipart.
 "jagged_lines"
 
+#' 3D jagged line with Z-dimension for smoothing
+#'
+#' Spatial lines in `sf` format for smoothing in three dimensions. There are
+#' examples of open and closed loops
+#'
+#' @format An `sf` object with 9 features and 3 attribute:
+#'   - **`type`**: character; the geometry, i.e. "polygon" or "line".
+#'   - **`closed`**: logical; whether the line forms a closed loop or not.
+#'   - **`multipart`**: logical; whether the feature is single or multipart.
+"jagged_lines_3d"
+
 #' Simulated raster for polygonizing and smoothing
 #'
 #' One of the primary applications of this package is for smoothing polygons