+++ R/smooth.r
@@ -153,9 +153,6 @@ smooth.sfg <- function(x, method = c("chaikin", "ksmooth", "spline", "densify"),
 smooth.sfc <- function(x, method = c("chaikin", "ksmooth", "spline", "densify"),
                        ...) {
   method <- match.arg(method)
-  if (!is.null(sf::st_z_range(x))) {
-    stop("Input geometry has unsupported Z dimension")
-  }
 
   for (i in seq_along(x)) {
     x[[i]] <- smooth(x[[i]], method = method, ...)
@@ -167,10 +164,6 @@ smooth.sfc <- function(x, method = c("chaikin", "ksmooth", "spline", "densify"),
 smooth.sf <- function(x, method = c("chaikin", "ksmooth", "spline", "densify"),
                       ...) {
   method <- match.arg(method)
-  if (!is.null(sf::st_z_range(x))) {
-    stop("Input geometry has unsupported Z dimension")
-  }
-
   sf::st_geometry(x) <- smooth(sf::st_geometry(x), method = method, ...)
   x
 }